Updating to quick-sip 0.2.0.
