+++ README.md
@@ -10,38 +10,39 @@
 > npm install -g generator-torso-brec
 
 ### Create a new project
-> cd path/to/project  
+> cd path/to/project
 > yo torso-brec
 
 In your current directory, the generator will produce the following files:
 
-  ├─ .gitignore  
-  ├─ package.json  
-  ├─ gulpfile.js  
-  ├─ dist  
-  └─ app  
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;     ├─ main.js  
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;     ├─ app.scss  
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;     ├─ index.html  
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;     └─ home  
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;        ├─ homeView.js  
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;        ├─ home-template.hbs  
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;        └─ _home.scss  
+    ├─ .gitignore
+    ├─ package.json
+    ├─ gulpfile.js
+    ├─ dist
+    └─ app
+        ├─ app.scss
+        ├─ app.js
+        ├─ router.js
+        ├─ index.html
+        └─ home
+            ├─ homeView.js
+            ├─ home-template.hbs
+            └─ _home.scss
 
 ## Pod Generator
 
-When creating an application, generating a new feature typically requires creating a view, template for that view, and sometimes special styling for that view as well. Pods are directories filled with these feature-specific files. Use the pod sub generator to rapidly create a pod pre-filled with a view, template, and scss file.  
+When creating an application, generating a new feature typically requires creating a view, template for that view, and sometimes special styling for that view as well. Pods are directories filled with these feature-specific files. Use the pod sub generator to rapidly create a pod pre-filled with a view, template, and scss file.
 
 ### Create a new pod
-> cd path/to/project/app  
-> yo torso-brec:pod foo  
+> cd path/to/project/app
+> yo torso-brec:pod foo
 
-This will create the following files:  
+This will create the following files:
 
-└─ foo  
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;     ├─ fooView.js  
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;     ├─ foo-template.hbs  
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;     └─ _foo.scss  
+    └─ foo
+       ├─ fooView.js
+       ├─ foo-template.hbs
+       └─ _foo.scss
 
 If an app.scss file exists, it will add: @import 'foo/foo'; to that file
